Copy paste query, types from library to dbcore
