+++ beets/ui/__init__.py
@@ -477,14 +477,6 @@ def human_seconds(interval):
     return f"{interval:3.1f} {suffix}s"
 
 
-def human_seconds_short(interval):
-    """Formats a number of seconds as a short human-readable M:SS
-    string.
-    """
-    interval = int(interval)
-    return "%i:%02i" % (interval // 60, interval % 60)
-
-
 # Colorization.
 
 # ANSI terminal colorization code heavily inspired by pygments:
