+++ beetsplug/metasync/itunes.py
@@ -26,7 +26,6 @@ from confuse import ConfigValueError
 
 from beets import util
 from beets.dbcore import types
-from beets.library import DateType
 from beets.util import bytestring_path, syspath
 from beetsplug.metasync import MetaSource
 
@@ -63,9 +62,9 @@ class Itunes(MetaSource):
         "itunes_rating": types.INTEGER,  # 0..100 scale
         "itunes_playcount": types.INTEGER,
         "itunes_skipcount": types.INTEGER,
-        "itunes_lastplayed": DateType(),
-        "itunes_lastskipped": DateType(),
-        "itunes_dateadded": DateType(),
+        "itunes_lastplayed": types.DATE,
+        "itunes_lastskipped": types.DATE,
+        "itunes_dateadded": types.DATE,
     }
 
     def __init__(self, config, log):
