+++ test/test_library.py
@@ -19,7 +19,6 @@ import os.path
 import re
 import shutil
 import stat
-import time
 import unicodedata
 import unittest
 from unittest.mock import patch
@@ -1320,56 +1319,3 @@ class ParseQueryTest(unittest.TestCase):
     def test_parse_bytes(self):
         with pytest.raises(AssertionError):
             beets.library.parse_query_string(b"query", None)
-
-
-class LibraryFieldTypesTest(unittest.TestCase):
-    """Test format() and parse() for library-specific field types"""
-
-    def test_datetype(self):
-        t = beets.library.DateType()
-
-        # format
-        time_format = beets.config["time_format"].as_str()
-        time_local = time.strftime(time_format, time.localtime(123456789))
-        assert time_local == t.format(123456789)
-        # parse
-        assert 123456789.0 == t.parse(time_local)
-        assert 123456789.0 == t.parse("123456789.0")
-        assert t.null == t.parse("not123456789.0")
-        assert t.null == t.parse("1973-11-29")
-
-    def test_pathtype(self):
-        t = beets.library.PathType()
-
-        # format
-        assert "/tmp" == t.format("/tmp")
-        assert "/tmp/\xe4lbum" == t.format("/tmp/\u00e4lbum")
-        # parse
-        assert np(b"/tmp") == t.parse("/tmp")
-        assert np(b"/tmp/\xc3\xa4lbum") == t.parse("/tmp/\u00e4lbum/")
-
-    def test_musicalkey(self):
-        t = beets.library.MusicalKey()
-
-        # parse
-        assert "C#m" == t.parse("c#m")
-        assert "Gm" == t.parse("g   minor")
-        assert "Not c#m" == t.parse("not C#m")
-
-    def test_durationtype(self):
-        t = beets.library.DurationType()
-
-        # format
-        assert "1:01" == t.format(61.23)
-        assert "60:01" == t.format(3601.23)
-        assert "0:00" == t.format(None)
-        # parse
-        assert 61.0 == t.parse("1:01")
-        assert 61.23 == t.parse("61.23")
-        assert 3601.0 == t.parse("60:01")
-        assert t.null == t.parse("1:00:01")
-        assert t.null == t.parse("not61.23")
-        # config format_raw_length
-        beets.config["format_raw_length"] = True
-        assert 61.23 == t.format(61.23)
-        assert 3601.23 == t.format(3601.23)
